Migrate back to layout and look ahead scope for simplicity
