+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/HomeScreen.kt
@@ -428,13 +428,14 @@ fun BoxScope.Success(
         Screen.Resize -> {
             ResizeScreen(
                 modifier = modifier,
-                currentPage = currentPage,
                 rows = userData.homeSettings.rows,
                 columns = userData.homeSettings.columns,
                 dockRows = userData.homeSettings.dockRows,
                 dockColumns = userData.homeSettings.dockColumns,
-                gridItems = gridItems,
+                gridItems = gridItems[currentPage],
                 gridItemLayoutInfo = gridItemSource?.gridItemLayoutInfo,
+                rootWidth = rootWidth,
+                rootHeight = rootHeight,
                 dockHeight = userData.homeSettings.dockHeight,
                 dockGridItems = dockGridItems,
                 textColor = userData.homeSettings.textColor,
@@ -463,6 +464,7 @@ fun BoxScope.Success(
                 dragIntOffset = dragIntOffset,
                 drag = drag,
                 rootWidth = rootWidth,
+                rootHeight = rootHeight,
                 dockHeight = userData.homeSettings.dockHeight,
                 movedCurrentPage = movedCurrentPage,
                 onSaveEditPage = onResetGridCache,
